+++ .gitattributes
@@ -1,0 +1,4 @@
+/.git*            export-ignore
+/phpunit.xml.dist export-ignore
+/README.md        export-ignore
+/tests            export-ignore
