--- .travis.yml
@@ -1,21 +1,0 @@
-language: php
-
-sudo: false
-
-cache:
-    directories:
-        - $HOME/.composer/cache
-
-php:
-    - 7.2
-    - 7.3
-    - 7.4
-
-matrix:
-    fast_finish: true
-
-install: composer install --prefer-dist --no-interaction
-
-before_script: ./vendor/bin/simple-phpunit install
-
-script: ./vendor/bin/simple-phpunit
