+++ composer.json
@@ -20,23 +20,23 @@
         }
     ],
     "require": {
-        "php": ">=7.2.5",
+        "php": "^8.0",
         "ext-curl": "*",
         "ext-intl": "*",
         "ext-json": "*",
         "ext-simplexml": "*",
-        "symfony/config": "^4.0|^5.0",
-        "symfony/console": "^4.0|^5.0",
-        "symfony/dependency-injection": "^4.0|^5.0",
-        "symfony/framework-bundle": "^4.0|^5.0",
-        "symfony/validator": "^4.0|^5.0",
-        "twig/twig": "^v2.14"
+        "symfony/config": "^5.0|^6.0",
+        "symfony/console": "^5.0|^6.0",
+        "symfony/dependency-injection": "^5.0|^6.0",
+        "symfony/framework-bundle": "^5.0|^6.0",
+        "symfony/validator": "^5.0|^6.0",
+        "twig/twig": "^v2.15|^3.0"
     },
     "require-dev": {
         "doctrine/data-fixtures": "^1.5",
         "doctrine/doctrine-bundle": "^1.8|^2.0",
-        "doctrine/orm": "^2.7",
-        "symfony/phpunit-bridge": "^4.3|^5.0"
+        "doctrine/orm": "^2.7|^3.0",
+        "symfony/phpunit-bridge": "^7.0"
     },
     "config": {
         "preferred-install": "dist",
